+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 ### Ported from [kamilsk/cilex-service-providers](https://github.com/kamilsk/CilexServiceProviders)
 - classes
   - `OctoLab\Cilex\ServiceProvider\ConfigServiceProvider`
+  - `OctoLab\Cilex\ServiceProvider\MonologServiceProvider`